+++ pkg/validation/app.go
@@ -2,6 +2,7 @@ package validation
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
 	"github.com/giantswarm/apiextensions/v3/pkg/label"
@@ -22,11 +23,6 @@ const (
 func (v *Validator) ValidateApp(ctx context.Context, app v1alpha1.App) (bool, error) {
 	var err error
 
-	err = v.validateLabels(ctx, app)
-	if err != nil {
-		return false, microerror.Mask(err)
-	}
-
 	err = v.validateCatalog(ctx, app)
 	if err != nil {
 		return false, microerror.Mask(err)
@@ -38,6 +34,16 @@ func (v *Validator) ValidateApp(ctx context.Context, app v1alpha1.App) (bool, er
 	}
 
 	err = v.validateKubeConfig(ctx, app)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
+	err = v.validateLabels(ctx, app)
+	if err != nil {
+		return false, microerror.Mask(err)
+	}
+
+	err = v.validateMetadataConstraints(ctx, app)
 	if err != nil {
 		return false, microerror.Mask(err)
 	}
@@ -97,14 +103,6 @@ func (v *Validator) validateConfig(ctx context.Context, cr v1alpha1.App) error {
 	return nil
 }
 
-func (v *Validator) validateLabels(ctx context.Context, cr v1alpha1.App) error {
-	if key.VersionLabel(cr) == "" {
-		return microerror.Maskf(validationError, labelNotFoundTemplate, label.AppOperatorVersion)
-	}
-
-	return nil
-}
-
 func (v *Validator) validateKubeConfig(ctx context.Context, cr v1alpha1.App) error {
 	if !key.InCluster(cr) {
 		ns := key.KubeConfigSecretNamespace(cr)
@@ -119,6 +117,68 @@ func (v *Validator) validateKubeConfig(ctx context.Context, cr v1alpha1.App) err
 			return microerror.Maskf(kubeConfigNotFoundError, resourceNotFoundTemplate, "kubeconfig secret", key.KubeConfigSecretName(cr), ns)
 		} else if err != nil {
 			return microerror.Mask(err)
+		}
+	}
+
+	return nil
+}
+
+func (v *Validator) validateLabels(ctx context.Context, cr v1alpha1.App) error {
+	if key.VersionLabel(cr) == "" {
+		return microerror.Maskf(validationError, labelNotFoundTemplate, label.AppOperatorVersion)
+	}
+
+	return nil
+}
+
+func (v *Validator) validateMetadataConstraints(ctx context.Context, cr v1alpha1.App) error {
+	name := key.AppCatalogEntryName(key.CatalogName(cr), key.AppName(cr), key.Version(cr))
+
+	entry, err := v.g8sClient.ApplicationV1alpha1().AppCatalogEntries(metav1.NamespaceDefault).Get(ctx, name, metav1.GetOptions{})
+	if apierrors.IsNotFound(err) {
+		v.logger.Debugf(ctx, "appcatalogentry %#q not found, skipping metadata validation", name)
+		return nil
+	} else if err != nil {
+		return microerror.Mask(err)
+	}
+
+	if entry.Spec.Restrictions == nil {
+		// no-op
+		return nil
+	}
+
+	if entry.Spec.Restrictions.FixedNamespace != "" {
+		if entry.Spec.Restrictions.FixedNamespace != cr.Spec.Namespace {
+			return microerror.Maskf(validationError, "app %#q can only be installed in namespace %#q only, not %#q",
+				cr.Spec.Name, entry.Spec.Restrictions.FixedNamespace, cr.Spec.Namespace)
+		}
+	}
+
+	var apps []v1alpha1.App
+	if entry.Spec.Restrictions.ClusterSingleton || entry.Spec.Restrictions.NamespaceSingleton {
+		lo := metav1.ListOptions{
+			FieldSelector: fmt.Sprintf("metadata.name!=%s", cr.Name),
+		}
+		appList, err := v.g8sClient.ApplicationV1alpha1().Apps(cr.Namespace).List(ctx, lo)
+		if err != nil {
+			return microerror.Mask(err)
+		}
+
+		apps = appList.Items
+	}
+
+	for _, app := range apps {
+		if app.Spec.Name == cr.Spec.Name {
+			if entry.Spec.Restrictions.ClusterSingleton {
+				return microerror.Maskf(validationError, "app %#q can only be installed once in cluster %#q",
+					cr.Spec.Name, cr.Namespace)
+			}
+			if entry.Spec.Restrictions.NamespaceSingleton {
+				if app.Spec.Namespace == cr.Spec.Namespace {
+					return microerror.Maskf(validationError, "app %#q can only be installed only once in namespace %#q",
+						cr.Spec.Name, key.Namespace(cr))
+				}
+			}
 		}
 	}